Implement the token exchange
main will call db_ops to request a token from database.
If the etsy onboarding flag is true and there is a token in the databse with an active time to live (at least 10mins) then that token is returned,
otherwise db_ops will call etsy_ops to request a token from the etsy api.
If this is the first time a token is being requested then the reference code form the db is used, else the refresh token is sent